Add nevalf_max as parameter:
This option is quite redundant with the max number of steps proposed by
the engine. However, as this number is not an input of artn(), it is not
possible to guess into artn when the last possible step is reached.
Using nevalf_max, when this number is smaller than the one of the engine
(if it exists) this permits to stop correctly the search when this max
number of force evalf is reached, i.e., by using lconv=.true. to recover
the starting structure and not to stop with the structure in the middle
of a search.