add optional argument to "add_trial_compiler" functions:
there is a bug when CXX is defined in environment, but is different than
CXX prescribed by lammps makefile. Solution here is a bit general, to add an
optional argument to add_trial_compiler, which can force writing to beginning of
list, even if the same compiler is already there.