create ARTN_RANDOM_NUMBER():
Separate the number generator used in artn from the global random_number()
generator. The idea is to use the `random_seed()` to switch between states of
RNG at each call to `ARTN_RANDOM_NUMBER()`, to ensure we do not perturb the
global state.
The steps:
 - save external_state   :`call random_seed(get=other_state)`
 - put saved artn_state  :`call random_seed(put=my_state)`
 - get random number z   :`call random_number(z)`
 - save new artn_state   :`call random_seed(get=my_state)`
 - re-put external state :`call random_seed(put=other_state)`
Like this RNG state of the engine is effectively not touched, and we
keep progressing the internal RNG state of artn.